requirements.txt
Flask
python-dotenv
requests
PyGithub
gitpython
openai
